fix: fix compile errors

+++ Tests/UnitTests/Math/MatrixTests.cc
@@ -6,42 +6,30 @@ using namespace Cowdia;
 
 TEST_CASE("[Math/Matrix] - Basic Arithmetic Operators")
 {
-	constexpr float eps = 1e-5;
+    constexpr float eps = 1e-5f;
 
-    Math::Matrix A{
-        1.0f,  2.0f,  3.0f,  4.0f, 
-        5.0f,  6.0f,  7.0f,  8.0f,
-        9.0f,  10.0f, 11.0f, 12.0f, 
-        13.0f, 14.0f, 15.0f, 16.0f
-    };
+    Math::Matrix A{ 1.0f, 2.0f,  3.0f,  4.0f,  5.0f,  6.0f,  7.0f,  8.0f,
+                    9.0f, 10.0f, 11.0f, 12.0f, 13.0f, 14.0f, 15.0f, 16.0f };
 
-    Math::Matrix B{
-        3.0f,  4.0f,  5.0f,  6.0f, 
-        7.0f,  8.0f,  9.0f,  10.0f, 
-        11.0f, 12.0f, 13.0f, 14.0f, 
-        15.0f, 16.0f, 17.0f, 18.0f
-    };
+    Math::Matrix B{ 3.0f,  4.0f,  5.0f,  6.0f,  7.0f,  8.0f,  9.0f,  10.0f,
+                    11.0f, 12.0f, 13.0f, 14.0f, 15.0f, 16.0f, 17.0f, 18.0f };
 
-    Math::Matrix C{
-        1.0f, 1.0f, 1.0f, 1.0f, 
-        1.0f, 1.0f, 1.0f, 1.0f, 
-        1.0f, 1.0f, 1.0f, 1.0f, 
-        1.0f, 1.0f, 1.0f, 1.0f
-    };
+    Math::Matrix C{ 1.0f, 1.0f, 1.0f, 1.0f, 1.0f, 1.0f, 1.0f, 1.0f,
+                    1.0f, 1.0f, 1.0f, 1.0f, 1.0f, 1.0f, 1.0f, 1.0f };
 
-    Math::Matrix D{ 
-        1.0f, 1.0f, 1.0f, 1.0f, 
-        1.0f, 1.0f, 1.0f, 1.0f, 
-        1.0f, 1.0f, 1.0f, 1.0f, 
-        1.0f, 1.0f, 1.0f, 1.0f 
-    };
+    Math::Matrix D{ 1.0f, 1.0f, 1.0f, 1.0f, 1.0f, 1.0f, 1.0f, 1.0f,
+                    1.0f, 1.0f, 1.0f, 1.0f, 1.0f, 1.0f, 1.0f, 1.0f };
 
     for (std::size_t i = 0; i < Math::Matrix::MAT_SIZE; ++i)
     {
         for (std::size_t j = 0; j < Math::Matrix::MAT_SIZE; ++j)
         {
-            REQUIRE(A.At(i, j) == doctest::Approx(i * 4 + j + 1).epsilon(eps));
-            REQUIRE(B.At(i, j) == doctest::Approx(i * 4 + j + 3).epsilon(eps));
+            REQUIRE(A.At(i, j) ==
+                    doctest::Approx(static_cast<float>(i * 4 + j + 1))
+                        .epsilon(eps));
+            REQUIRE(B.At(i, j) ==
+                    doctest::Approx(static_cast<float>(i * 4 + j + 3))
+                        .epsilon(eps));
         }
     }
 
@@ -52,7 +40,8 @@ TEST_CASE("[Math/Matrix] - Basic Arithmetic Operators")
         for (std::size_t j = 0; j < Math::Matrix::MAT_SIZE; ++j)
         {
             REQUIRE(AplusB.At(i, j) ==
-                    doctest::Approx(i * 8 + j * 2 + 4).epsilon(eps));
+                    doctest::Approx(static_cast<float>(i * 8 + j * 2 + 4))
+                        .epsilon(eps));
         }
     }
 
@@ -65,7 +54,7 @@ TEST_CASE("[Math/Matrix] - Basic Arithmetic Operators")
         }
     }
 
-	Math::Matrix CmulD = C * D;
+    Math::Matrix CmulD = C * D;
     for (std::size_t i = 0; i < Math::Matrix::MAT_SIZE; ++i)
     {
         for (std::size_t j = 0; j < Math::Matrix::MAT_SIZE; ++j)
@@ -74,23 +63,25 @@ TEST_CASE("[Math/Matrix] - Basic Arithmetic Operators")
         }
     }
 
-	Math::Matrix Aplus2 = A + 2;
+    Math::Matrix Aplus2 = A + 2;
     for (std::size_t i = 0; i < Math::Matrix::MAT_SIZE; ++i)
     {
         for (std::size_t j = 0; j < Math::Matrix::MAT_SIZE; ++j)
         {
             REQUIRE(Aplus2.At(i, j) ==
-                    doctest::Approx(i * 4 + j + 3).epsilon(eps));
+                    doctest::Approx(static_cast<float>(i * 4 + j + 3))
+                        .epsilon(eps));
         }
     }
 
-	Math::Matrix Aminus2 = A - 2;
+    Math::Matrix Aminus2 = A - 2;
     for (std::size_t i = 0; i < Math::Matrix::MAT_SIZE; ++i)
     {
         for (std::size_t j = 0; j < Math::Matrix::MAT_SIZE; ++j)
         {
             REQUIRE(Aminus2.At(i, j) ==
-                    doctest::Approx(i * 4 + j - 1).epsilon(eps));
+                    doctest::Approx(static_cast<float>(i * 4 + j - 1))
+                        .epsilon(eps));
         }
     }
 
@@ -100,7 +91,8 @@ TEST_CASE("[Math/Matrix] - Basic Arithmetic Operators")
         for (std::size_t j = 0; j < Math::Matrix::MAT_SIZE; ++j)
         {
             REQUIRE(Amul.At(i, j) ==
-                    doctest::Approx(i * 12 + j * 3 + 3).epsilon(eps));
+                    doctest::Approx(static_cast<float>(i * 12 + j * 3 + 3))
+                        .epsilon(eps));
         }
     }
 
@@ -110,47 +102,52 @@ TEST_CASE("[Math/Matrix] - Basic Arithmetic Operators")
         for (std::size_t j = 0; j < Math::Matrix::MAT_SIZE; ++j)
         {
             REQUIRE(Adiv.At(i, j) ==
-                    doctest::Approx(i * 2 + (j + 1) / 2).epsilon(eps));
+                    doctest::Approx(static_cast<float>(i * 2 + (j + 1) / 2))
+                        .epsilon(eps));
         }
     }
 
-	Math::Matrix Atrans = A.T();
+    Math::Matrix Atrans = A.T();
     for (std::size_t i = 0; i < Math::Matrix::MAT_SIZE; ++i)
     {
         for (std::size_t j = 0; j < Math::Matrix::MAT_SIZE; ++j)
         {
             REQUIRE(Atrans.At(j, i) ==
-                    doctest::Approx(i * 4 + j + 1).epsilon(eps));
+                    doctest::Approx(static_cast<float>(i * 4 + j + 1))
+                        .epsilon(eps));
         }
     }
 
-	Math::Matrix TwoplusA = 2 + A;
+    Math::Matrix TwoplusA = 2 + A;
     for (std::size_t i = 0; i < Math::Matrix::MAT_SIZE; ++i)
     {
         for (std::size_t j = 0; j < Math::Matrix::MAT_SIZE; ++j)
         {
             REQUIRE(TwoplusA.At(i, j) ==
-                    doctest::Approx(i * 4 + j + 3).epsilon(eps));
+                    doctest::Approx(static_cast<float>(i * 4 + j + 3))
+                        .epsilon(eps));
         }
     }
 
-	Math::Matrix TwominusA = 2 - A;
+    Math::Matrix TwominusA = 2 - A;
     for (std::size_t i = 0; i < Math::Matrix::MAT_SIZE; ++i)
     {
         for (std::size_t j = 0; j < Math::Matrix::MAT_SIZE; ++j)
         {
             REQUIRE(TwominusA.At(i, j) ==
-                    doctest::Approx(- i * 4 - j + 1).epsilon(eps));
+                    doctest::Approx(static_cast<float>(-static_cast<int>(i) * 4 - j + 1))
+                        .epsilon(eps));
         }
     }
 
-	Math::Matrix Threemul = A * 3;
+    Math::Matrix Threemul = A * 3;
     for (std::size_t i = 0; i < Math::Matrix::MAT_SIZE; ++i)
     {
         for (std::size_t j = 0; j < Math::Matrix::MAT_SIZE; ++j)
         {
             REQUIRE(Threemul.At(i, j) ==
-                    doctest::Approx(i * 12 + j * 3 + 3).epsilon(eps));
+                    doctest::Approx(static_cast<float>(i * 12 + j * 3 + 3))
+                        .epsilon(eps));
         }
     }
 }
